+++ CHANGELOG.md
@@ -9,11 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **GM-71** - Best Courses component.
+### Fixed
+
+## [0.6.0] - 2022-06-29
+
+### Added
+- **GM-89** - Product Shelf with menu selector.
 - **GM-32** - Main Banner component.
 
 ### Fixed
 
 - **GM-267** - Video section styles.
+- **GM-294** - Fix shelf title typo.
 
 ## [0.5.0] - 2022-06-23